requirements.txt
ultralytics==8.0.47
roboflow==0.2.31
python-dotenv==1.0.0